Fix #721, Fix #729: Web view top offset now properly (always) accounts for tab bar.
webViewContainer was ending up with multiple, conflicting constraints, some accounted for the tab bar and others did not. This ended up in a weird, unpredictable place with layout for the top offset. webViewContainer constraints are now completely replaced when adjustments are required. This should be refactored to only use the offset adjustment and not replace the constraints entirely (although this is a good security to have in place).